+++ GHEtool/VariableClasses/__init__.py
@@ -4,5 +4,5 @@ from .LoadData import *
 from .PipeData import *
 from .CustomGFunction import CustomGFunction, load_custom_gfunction, _timeValues
 from .GFunction import GFunction, FIFO
-from .SizingSetup import SizingSetup
+from .CalculationSetup import CalculationSetup
 from .Borehole import Borehole